chore: rename refresh columns / add doc / add auto run
