+++ app/tasks/maintenance/t20250115_refresh_columns_id_in_procedure_presentation_task.rb
@@ -1,8 +1,13 @@
 # frozen_string_literal: true
 
+# This task run over all procedure_presentations to reload the columns and reserialize them using their current id
+# In order to be refreshed, the column must be unserialized so the method find_column in ColumnConcern must be adapted to work with the old and new id format
+# this task should be use with the task t20250115refreshColumnsIdInExportTask, t20250115refreshColumnsIdInExportTemplateTask, t20250115refreshColumnsIdInProcedurePresentationTask
 module Maintenance
-  class MigrateProcedurePresentationAddressAndLinkedColumnsTask < MaintenanceTasks::Task
+  class T20250115RefreshColumnsIdInProcedurePresentationTask < MaintenanceTasks::Task
     include RunnableOnDeployConcern
+
+    run_on_first_deploy
 
     def collection
       ProcedurePresentation.all
