+++ spec/tasks/maintenance/t20250115_refresh_columns_id_in_procedure_presentation_task_spec.rb
@@ -3,7 +3,7 @@
 require "rails_helper"
 
 module Maintenance
-  RSpec.describe MigrateProcedurePresentationAddressAndLinkedColumnsTask do
+  RSpec.describe T20250115RefreshColumnsIdInProcedurePresentationTask do
     def displayed_column_ids
       raw = ProcedurePresentation
         .connection
